Flink_server/requirements.txt
boto3
pillow
apache-flink==1.18.1
apache-flink-libraries==1.18.1

